+++ packages/br-validators/src/index.ts
@@ -2,3 +2,4 @@ export * from './cnpj.js';
 export * from './cpf.js';
 export * from './cep.js';
 export * from './placa.js';
+export * from './pis-pasep.js';